Issue with parameters for native query
Hey guys, tinkering with micronaut data hibernate, I identified an issue about bind query parameters. This issue happens when I'm binding a Type List, Object[] or NULL parameter. Basically, the routine breaks the parameter assignment when it's assigned to one of the listed ones.

E.g: If i set four parameters (two Lists and two Strings), the code - depending on the order of the parameters - passes those two Strings and after the first List is made, an error will be returned, not passing the second List.

Is this a bug? I'm not sure this is the correct behaviour.