Fix compilation flakiness
This commit fixes an accidental mutation of the static `PREFIXES` field which
was happening when sorting the list of prefixes in `AbstractPrefixPatternMethodMatcher`.

The consequence was the creation, sometimes, of incorrect patterns. For example,
the generated pattern could be `^((persist|insert|insert|store)(\S*?))$`
instead of `^((persist|insert|save|store)(\S*?))$` (notice the duplicate entry).

It also replaces the static array type with an unmodifiable list, as it's in
general not a good idea to have mutable datastructures as static fields.

Fixes #1400