Manually instantiate ObjectMapper
The `CosmosSerde` class internally uses Jackson to parse CosmosDB
documents. Before Micronaut 4, there was a high likelyhood that
Micronaut would use Jackson for serialization, and not Micronaut
Serde.

By moving to Micronaut 4, we can't be sure that there will be a
bean of Jackson's `ObjectMapper` type, so we create one for this
class only.