Fix: Oracle ClientInfo should not always be reset. (#3461)
Calling a @Connectable/@Transactional method from inside another one left the
connection without any client info set after the inner method returns. In code
that grabbed the Connection and used it, that would result in calls that lacked
client info.

A better approach is to properly save and restore the client info.